refactor(tmux): split interface into multiple files
- move struct definitions and implementations to `src/tmux/session.rs`.
- move logic for interacting with tmux to `src/tmux/interface.rs`.
- add module exports in `src/tmux/mod.rs`.
